# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Session {
  id: number;
  title: string;
  model: string;
  createdAt: string;
  updatedAt: string;
}

export interface SessionInput {
  title: string;
  model?: string;
}

export interface SessionPreview {
  id: number;
  title: string;
  model: string;
  /** @nullable */
  lastMessage?: string | null;
  messageCount: number;
  updatedAt: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
  system: 'system',
} as const;

export interface Message {
  id: number;
  sessionId: number;
  role: MessageRole;
  content: string;
  /** @nullable */
  toolResults?: string | null;
  createdAt: string;
}

export interface SessionWithMessages {
  id: number;
  title: string;
  model: string;
  messages: Message[];
  createdAt: string;
  updatedAt: string;
}

export interface ChatInput {
  sessionId: number;
  message: string;
  model?: string;
  webSearch?: boolean;
}

export interface Model {
  id: string;
  name: string;
  contextLength: number;
}

export type TemplateCategory = typeof TemplateCategory[keyof typeof TemplateCategory];


export const TemplateCategory = {
  nginx: 'nginx',
  systemd: 'systemd',
  docker: 'docker',
  deploy: 'deploy',
} as const;

export interface Template {
  id: string;
  name: string;
  description: string;
  category: TemplateCategory;
  language: string;
}

export interface TemplateDetail {
  id: string;
  name: string;
  description: string;
  category: string;
  language: string;
  content: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Note {
  id: number;
  title: string;
  body?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateNote {
  /** @minLength 1 */
  title: string;
  body?: string | null;
}

export interface UpdateNote {
  /** @minLength 1 */
  title?: string;
  body?: string | null;
}

export interface ValidationError {
  error: string;
}

export interface NotFoundError {
  error: string;
}

export interface AuthUser {
  id: string;
  email?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  profileImageUrl?: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export interface LogoutSuccess {
  success: boolean;
}

export interface ErrorEnvelope {
  error: string;
}

export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type LogoutBrowserSessionParams = {
returnTo?: string;
};