# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/chat:
    post:
      operationId: sendChat
      tags: [ai]
      summary: Send a chat message and get a streaming AI response
      description: Returns a non-streaming summary; actual streaming uses /api/ai/stream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatInput"
      responses:
        "200":
          description: AI reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Message:
      type: object
      properties:
        id:
          type: number
        sessionId:
          type: number
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        toolResults:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, sessionId, role, content, createdAt]
    ChatInput:
      type: object
      properties:
        sessionId:
          type: number
        message:
          type: string
        model:
          type: string
        webSearch:
          type: boolean
      required: [sessionId, message]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...